Latest changes to V3 firmware - still WIP

+++ Firmware_V3/lib/MTP/Storage.cpp
@@ -298,14 +298,20 @@ void mtp_lock_storage(bool lock) {}
 void MTPStorage_SD::removeFile(uint32_t store, char *file)
 { 
   char tname[MAX_FILENAME_LEN];
-  
   File f1=sd_open(store,file,0);
-  File f2;
-  while(f2=f1.openNextFile())
-  { sprintf(tname,"%s/%s",file,f2.name());
-    if(f2.isDirectory()) removeFile(store,tname); else sd_remove(store,tname);
+  if(f1.isDirectory())
+  {
+    File f2;
+    while(f2=f1.openNextFile())
+    { sprintf(tname,"%s/%s",file,f2.name());
+      if(f2.isDirectory()) removeFile(store,tname); else sd_remove(store,tname);
+    }
+    sd_rmdir(store,file);
   }
-  sd_rmdir(store,file);
+  else
+  {
+    sd_remove(store,file);
+  }
 }
 
   bool MTPStorage_SD::DeleteObject(uint32_t object)
